+++ Changelog.md
@@ -6,6 +6,7 @@ MacInfoPkg changelog
 - Added MacBookPro15,4 model codes
 - Added MacBookPro15,3 model codes
 - Added MacBookAir8,2 model codes
+- Added macserial basic other product recognition
 
 #### v2.0.6
 - Initial model fillup (thx to all participants)